NSTableColumn: keep the width the column was given
-width, -minWidth and -maxWidth return a CGFloat, but the column held the
three in float ivars, so on a 64 bit system every width was rounded to
float precision on the way in and came back changed: after
[column setWidth: 123.456789] the column reported 123.45678710937501.

The ivars are CGFloat now.  The archive format is unchanged: the three are
still encoded and decoded as float, through a local variable, so a column
written by an earlier version still reads.

+++ Headers/AppKit/NSTableColumn.h
@@ -63,9 +63,9 @@ APPKIT_EXPORT_CLASS
 {
   id _identifier;
   NSTableView *_tableView;
-  float _width;
-  float _min_width;
-  float _max_width;
+  CGFloat _width;
+  CGFloat _min_width;
+  CGFloat _max_width;
   NSUInteger _resizing_mask;
   BOOL _is_resizable;
   BOOL _is_editable;
